requirements.txt
spotipy
yt-dlp
python-dotenv
tqdm
gradio
